#22 rename component
