+++ src/router/index.ts
@@ -3,7 +3,7 @@ import { Routes } from '@/router/types';
 import StartPage from '@/views/StartPage.vue';
 import RegisterCooperation from '@/views/RegisterCooperation.vue';
 import UserLogin from '@/views/UserLogin.vue';
-import UserRegistration from '@/views/UserRegistration.vue';
+import RegisterUser from '@/views/RegisterUser.vue';
 
 const routes: RouteRecordRaw[] = [
   // initial route; should be customized or removed
@@ -21,7 +21,7 @@ const routes: RouteRecordRaw[] = [
   },
   {
     path: Routes.RegisterUser,
-    component: UserRegistration,
+    component: RegisterUser,
   },
 ];
